Minor updates to the MD documentation files
I've made a bunch of minor updates to the markdown files based on observations from reading them for the first time.
Most of them are trivial updates, but check annotations.md in particular as was unsure about the third bullet point of section "A Suitepath can be provided in three ways".