I made the change for that the guide don't say that ut3 has the last develop version, if not that It has the version of utPLSQL of your current branch in your locally repository.
Another change I made was that the last line in the "That's it" section say
this: At any time, if you need to refresh your database, the below scripts might
be helpful.
But next to that section is the "Running unit tests" section, they are not the
script that was being talked about, because they are in the secction following
"Running unit tests". So I changed the order